requirements.txt
bpython
django
djangorestframework
ecdsa
pbkdf2
requests[security]
psycopg2
python-bitcoinrpc
